+++ core/data/src/test/java/com/sidgowda/pawcalc/data/dogs/model/DogWeightStringFormatterTest.kt
@@ -1,4 +1,4 @@
-package com.sidgowda.pawcalc.data.dogs
+package com.sidgowda.pawcalc.data.dogs.model
 
 import com.sidgowda.pawcalc.data.dogs.model.formattedToString
 import io.kotest.matchers.shouldBe
