+++ core/data/src/test/java/com/sidgowda/pawcalc/data/settings/repo/SettingsRepoTest.kt
@@ -1,4 +1,4 @@
-package com.sidgowda.pawcalc.data.settings
+package com.sidgowda.pawcalc.data.settings.repo
 
 import app.cash.turbine.test
 import com.sidgowda.pawcalc.common.settings.DateFormat
@@ -6,8 +6,6 @@ import com.sidgowda.pawcalc.common.settings.ThemeFormat
 import com.sidgowda.pawcalc.common.settings.WeightFormat
 import com.sidgowda.pawcalc.data.fakes.FakeSettingsDataSource
 import com.sidgowda.pawcalc.data.settings.model.Settings
-import com.sidgowda.pawcalc.data.settings.repo.SettingsRepo
-import com.sidgowda.pawcalc.data.settings.repo.SettingsRepoImpl
 import junit.framework.TestCase.assertEquals
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.Flow
