+++ feature/dogdetails/src/test/java/com/sidgowda/pawcalc/dogdetails/ui/CircleUtilsTest.kt
@@ -1,0 +1,4 @@
+package com.sidgowda.pawcalc.dogdetails.ui
+
+class CircleUtilsTest {
+}
