+++ feature/doglist/src/test/java/com/sidgowda/pawcalc/doglist/ui/DogListViewModelTest.kt
@@ -1,4 +1,4 @@
-package com.sidgowda.pawcalc.doglist
+package com.sidgowda.pawcalc.doglist.ui
 
 import android.net.Uri
 import androidx.lifecycle.SavedStateHandle
@@ -14,10 +14,11 @@ import com.sidgowda.pawcalc.data.dogs.model.formattedToTwoDecimals
 import com.sidgowda.pawcalc.data.dogs.model.toNewWeight
 import com.sidgowda.pawcalc.data.dogs.repo.DogsRepo
 import com.sidgowda.pawcalc.data.onboarding.model.OnboardingState
+import com.sidgowda.pawcalc.doglist.FakeDogsDataSource
+import com.sidgowda.pawcalc.doglist.FakeDogsRepo
 import com.sidgowda.pawcalc.doglist.model.DogListEvent
 import com.sidgowda.pawcalc.doglist.model.DogListState
 import com.sidgowda.pawcalc.doglist.model.NavigateEvent
-import com.sidgowda.pawcalc.doglist.ui.DogListViewModel
 import com.sidgowda.pawcalc.domain.onboarding.GetOnboardingStateUseCase
 import com.sidgowda.pawcalc.test.MainDispatcherRule
 import io.kotest.matchers.collections.shouldContainExactly
