+++ feature/onboarding/src/test/java/com/sidgowda/pawcalc/onboarding/ui/OnboardingViewModelTest.kt
@@ -1,4 +1,4 @@
-package com.sidgowda.pawcalc.onboarding
+package com.sidgowda.pawcalc.onboarding.ui
 
 import com.sidgowda.pawcalc.domain.onboarding.SetUserOnboardedUseCase
 import com.sidgowda.pawcalc.onboarding.ui.OnboardingViewModel
